Added controllers .

+++ restaurant-data/src/main/java/com/app/resturant/model/Recipe.java
@@ -17,11 +17,11 @@ public class Recipe extends BaseEntity {
     private final String name;
     List<KitchenWare> neededKitchenWare;
     Map<Long, String> instructions;
-    Map<IngredientType, Long> neededIngredients;
+    Map<IngredientType, Integer> neededIngredients;
 
 
     @Builder
-    public Recipe(String name, Map<IngredientType, Long> neededIngredients, List<KitchenWare> neededKitchenWare, Map<Long, String> instructions) {
+    public Recipe(String name, Map<IngredientType, Integer> neededIngredients, List<KitchenWare> neededKitchenWare, Map<Long, String> instructions) {
         this.name = name;
         this.neededIngredients = neededIngredients;
         this.neededKitchenWare = neededKitchenWare;
