+++ restaurant-data/src/main/java/com/app/resturant/repositories/IngredientTypeRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.HashMap;
 
 public interface IngredientTypeRepository  extends CrudRepository<IngredientType,Long> {
-    HashMap<IngredientType,Integer> findByName(IngredientType ingredientType);
+    HashMap<IngredientType,Long> findByName(IngredientType ingredientType);
 }
